src/jvis/data/stacks/python-django/files/core/admin.py
"""Core admin configuration."""

from django.contrib import admin

from core.models import Item


@admin.register(Item)
class ItemAdmin(admin.ModelAdmin):
    list_display = ["name", "created_at", "updated_at"]
    search_fields = ["name", "description"]
    readonly_fields = ["id", "created_at", "updated_at"]
