src/jvis/data/stacks/python-django/files/requirements-dev.txt
# Development Dependencies
-r requirements.txt
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
ruff>=0.8.0
mypy>=1.10.0
django-stubs>=5.0.0
coverage>=7.4.0
